Microservices/services/requirements.txt
Django>=4.2
djangorestframework>=3.14
mysqlclient>=2.1
requests>=2.28
django-cors-headers>=4.0
